add create eth wallet in rest_api

+++ backend/rest_api/src/domain/value_objects/wallet/balance.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 from src.domain.exceptions.fields import NegativeValueException
 from src.domain.value_objects.base import ValueObject
@@ -6,13 +7,13 @@ from src.domain.value_objects.base import ValueObject
 
 @dataclass(frozen=True, slots=True, repr=False)
 class Balance(ValueObject):
-    value: int
+    value: Decimal
 
     def __post_init__(self):
         super(Balance, self).__post_init__()
         self._validate_balance(self.value)
 
     @staticmethod
-    def _validate_balance(balance_value: int) -> None:
+    def _validate_balance(balance_value: Decimal) -> None:
         if balance_value < 0:
             raise NegativeValueException(field="balance")
